# Conflicts:
#	modules/cluster-management/src/main/java/org/apache/ignite/internal/cluster/management/ClusterManagementGroupManager.java
#	modules/cluster-management/src/test/java/org/apache/ignite/internal/cluster/management/ClusterManagementGroupManagerTest.java
#	modules/core/src/main/java/org/apache/ignite/internal/lang/IgniteSystemProperties.java
#	modules/raft/src/main/java/org/apache/ignite/internal/raft/server/impl/JraftServerImpl.java
#	modules/raft/src/main/java/org/apache/ignite/raft/jraft/core/NodeImpl.java
#	modules/raft/src/main/java/org/apache/ignite/raft/jraft/option/NodeOptions.java
#	modules/raft/src/main/java/org/apache/ignite/raft/jraft/storage/impl/LogManagerImpl.java
#	modules/raft/src/test/java/org/apache/ignite/disruptor/StripedDisruptorTest.java
#	modules/runner/src/integrationTest/java/org/apache/ignite/internal/benchmark/AbstractMultiNodeBenchmark.java
#	modules/runner/src/integrationTest/java/org/apache/ignite/internal/benchmark/UpsertKvBenchmark.java
#	modules/runner/src/integrationTest/java/org/apache/ignite/internal/runner/app/ItRaftCommandLeftInLogUntilRestartTest.java
#	modules/table/src/main/java/org/apache/ignite/internal/table/distributed/replicator/PartitionReplicaListener.java

+++ modules/storage-rocksdb/src/main/java/org/apache/ignite/internal/storage/rocksdb/RocksDbTableStorage.java
@@ -20,9 +20,11 @@ package org.apache.ignite.internal.storage.rocksdb;
 import static java.util.stream.Collectors.toList;
 import static org.apache.ignite.internal.storage.rocksdb.instance.SharedRocksDbInstance.DFLT_WRITE_OPTS;
 import static org.apache.ignite.internal.storage.util.StorageUtils.createMissingMvPartitionErrorMessage;
+import static org.apache.ignite.internal.storage.util.StorageUtils.throwExceptionDependingOnStorageState;
+import static org.apache.ignite.internal.storage.util.StorageUtils.transitionToClosedState;
+import static org.apache.ignite.internal.storage.util.StorageUtils.transitionToDestroyedState;
 import static org.apache.ignite.internal.util.CompletableFutures.nullCompletedFuture;
 import static org.apache.ignite.internal.util.IgniteUtils.closeAll;
-import static org.apache.ignite.internal.util.IgniteUtils.inBusyLock;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -30,21 +32,23 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
-import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicReference;
+import java.util.function.Supplier;
+import org.apache.ignite.internal.lang.IgniteStringFormatter;
 import org.apache.ignite.internal.storage.MvPartitionStorage;
 import org.apache.ignite.internal.storage.StorageException;
 import org.apache.ignite.internal.storage.StorageRebalanceException;
+import org.apache.ignite.internal.storage.engine.MvPartitionMeta;
 import org.apache.ignite.internal.storage.engine.MvTableStorage;
 import org.apache.ignite.internal.storage.engine.StorageTableDescriptor;
-import org.apache.ignite.internal.storage.index.HashIndexStorage;
 import org.apache.ignite.internal.storage.index.IndexStorage;
-import org.apache.ignite.internal.storage.index.SortedIndexStorage;
 import org.apache.ignite.internal.storage.index.StorageHashIndexDescriptor;
 import org.apache.ignite.internal.storage.index.StorageIndexDescriptorSupplier;
 import org.apache.ignite.internal.storage.index.StorageSortedIndexDescriptor;
 import org.apache.ignite.internal.storage.rocksdb.index.AbstractRocksDbIndexStorage;
 import org.apache.ignite.internal.storage.rocksdb.instance.SharedRocksDbInstance;
 import org.apache.ignite.internal.storage.util.MvPartitionStorages;
+import org.apache.ignite.internal.storage.util.StorageState;
 import org.apache.ignite.internal.util.IgniteSpinBusyLock;
 import org.jetbrains.annotations.Nullable;
 import org.rocksdb.ColumnFamilyHandle;
@@ -67,8 +71,7 @@ public class RocksDbTableStorage implements MvTableStorage {
     /** Busy lock to stop synchronously. */
     private final IgniteSpinBusyLock busyLock = new IgniteSpinBusyLock();
 
-    /** Prevents double stopping of the component. */
-    private final AtomicBoolean stopGuard = new AtomicBoolean();
+    private final AtomicReference<StorageState> state = new AtomicReference<>(StorageState.RUNNABLE);
 
     /** Table descriptor. */
     private final StorageTableDescriptor tableDescriptor;
@@ -97,7 +100,7 @@ public class RocksDbTableStorage implements MvTableStorage {
         try {
             indexes.recoverIndexes(indexDescriptorSupplier);
         } catch (RocksDBException e) {
-            throw new StorageException("Unable to recover indexes", e);
+            throw new IgniteRocksDbException("Unable to recover indexes", e);
         }
     }
 
@@ -150,11 +153,15 @@ public class RocksDbTableStorage implements MvTableStorage {
      * @param schedule {@code true} if {@link RocksDB#flush(FlushOptions)} should be explicitly triggered in the near future.
      */
     public CompletableFuture<Void> awaitFlush(boolean schedule) {
-        return inBusyLock(busyLock, () -> rocksDb.flusher.awaitFlush(schedule));
+        return busy(() -> rocksDb.flusher.awaitFlush(schedule));
     }
 
     private CompletableFuture<Void> stop(boolean destroy) {
-        if (!stopGuard.compareAndSet(false, true)) {
+        boolean transitionedToTerminalState = destroy
+                ? transitionToDestroyedState(state)
+                : transitionToClosedState(state, this::createStorageInfo);
+
+        if (!transitionedToTerminalState) {
             return nullCompletedFuture();
         }
 
@@ -203,7 +210,7 @@ public class RocksDbTableStorage implements MvTableStorage {
 
     @Override
     public CompletableFuture<MvPartitionStorage> createMvPartition(int partitionId) throws StorageException {
-        return inBusyLock(busyLock, () -> mvPartitionStorages.create(partitionId, partId -> {
+        return busy(() -> mvPartitionStorages.create(partitionId, partId -> {
             RocksDbMvPartitionStorage partition = new RocksDbMvPartitionStorage(this, partitionId);
 
             if (partition.lastAppliedIndex() == 0L) {
@@ -221,45 +228,57 @@ public class RocksDbTableStorage implements MvTableStorage {
 
     @Override
     public @Nullable RocksDbMvPartitionStorage getMvPartition(int partitionId) {
-        return inBusyLock(busyLock, () -> mvPartitionStorages.get(partitionId));
+        return busy(() -> mvPartitionStorages.get(partitionId));
     }
 
     @Override
     public CompletableFuture<Void> destroyPartition(int partitionId) {
-        return inBusyLock(busyLock, () -> mvPartitionStorages.destroy(partitionId, mvPartitionStorage -> {
-            try (WriteBatch writeBatch = new WriteBatch()) {
-                mvPartitionStorage.transitionToDestroyedState();
+        if (!busyLock.enterBusy()) {
+            return nullCompletedFuture();
+        }
 
-                // Operation to delete partition data should be fast, since we will write only the range of keys for deletion, and the
-                // RocksDB itself will then destroy the data on flush.
-                mvPartitionStorage.destroyData(writeBatch);
+        try {
+            return mvPartitionStorages.destroy(partitionId, mvPartitionStorage -> {
+                try (WriteBatch writeBatch = new WriteBatch()) {
+                    mvPartitionStorage.transitionToDestroyedState();
 
-                indexes.destroyAllIndexesForPartition(partitionId, writeBatch);
+                    // Operation to delete partition data should be fast, since we will write only the range of keys for deletion, and the
+                    // RocksDB itself will then destroy the data on flush.
+                    mvPartitionStorage.destroyData(writeBatch);
 
-                rocksDb.db.write(DFLT_WRITE_OPTS, writeBatch);
+                    indexes.destroyAllIndexesForPartition(partitionId, writeBatch);
 
-                return nullCompletedFuture();
-            } catch (RocksDBException e) {
-                throw new StorageException("Error when destroying storage: [{}]", e, mvPartitionStorages.createStorageInfo(partitionId));
-            }
-        }));
+                    rocksDb.db.write(DFLT_WRITE_OPTS, writeBatch);
+
+                    return nullCompletedFuture();
+                } catch (RocksDBException e) {
+                    throw new IgniteRocksDbException(
+                            String.format("Error when destroying storage: [%s]", mvPartitionStorages.createStorageInfo(partitionId)), e
+                    );
+                }
+            });
+        } finally {
+            busyLock.leaveBusy();
+        }
     }
 
     @Override
-    public SortedIndexStorage getOrCreateSortedIndex(int partitionId, StorageSortedIndexDescriptor indexDescriptor) {
-        return inBusyLock(busyLock, () -> {
-            checkPartitionExists(partitionId);
-
-            return indexes.getOrCreateSortedIndex(partitionId, indexDescriptor);
+    public void createSortedIndex(int partitionId, StorageSortedIndexDescriptor indexDescriptor) {
+        busy(() -> {
+            // TODO: IGNITE-24926 - invoke checkPartitionExists() instead to throw StorageException is partition is absent.
+            if (partitionExists(partitionId)) {
+                indexes.createSortedIndex(partitionId, indexDescriptor);
+            }
         });
     }
 
     @Override
-    public HashIndexStorage getOrCreateHashIndex(int partitionId, StorageHashIndexDescriptor indexDescriptor) {
-        return inBusyLock(busyLock, () -> {
-            checkPartitionExists(partitionId);
-
-            return indexes.getOrCreateHashIndex(partitionId, indexDescriptor);
+    public void createHashIndex(int partitionId, StorageHashIndexDescriptor indexDescriptor) {
+        busy(() -> {
+            // TODO: IGNITE-24926 - invoke checkPartitionExists() instead to throw StorageException is partition is absent.
+            if (partitionExists(partitionId)) {
+                indexes.createHashIndex(partitionId, indexDescriptor);
+            }
         });
     }
 
@@ -274,7 +293,7 @@ public class RocksDbTableStorage implements MvTableStorage {
 
             return nullCompletedFuture();
         } catch (RocksDBException e) {
-            throw new StorageException("Error when destroying index: {}", e, indexId);
+            throw new IgniteRocksDbException(String.format("Error when destroying index: %d", indexId), e);
         } finally {
             busyLock.leaveBusy();
         }
@@ -287,7 +306,7 @@ public class RocksDbTableStorage implements MvTableStorage {
 
     @Override
     public CompletableFuture<Void> startRebalancePartition(int partitionId) {
-        return inBusyLock(busyLock, () -> mvPartitionStorages.startRebalance(partitionId, mvPartitionStorage -> {
+        return busy(() -> mvPartitionStorages.startRebalance(partitionId, mvPartitionStorage -> {
             try (WriteBatch writeBatch = new WriteBatch()) {
                 mvPartitionStorage.startRebalance(writeBatch);
 
@@ -308,7 +327,7 @@ public class RocksDbTableStorage implements MvTableStorage {
 
     @Override
     public CompletableFuture<Void> abortRebalancePartition(int partitionId) {
-        return inBusyLock(busyLock, () -> mvPartitionStorages.abortRebalance(partitionId, mvPartitionStorage -> {
+        return busy(() -> mvPartitionStorages.abortRebalance(partitionId, mvPartitionStorage -> {
             try (WriteBatch writeBatch = new WriteBatch()) {
                 mvPartitionStorage.abortRebalance(writeBatch);
 
@@ -327,15 +346,10 @@ public class RocksDbTableStorage implements MvTableStorage {
     }
 
     @Override
-    public CompletableFuture<Void> finishRebalancePartition(
-            int partitionId,
-            long lastAppliedIndex,
-            long lastAppliedTerm,
-            byte[] groupConfig
-    ) {
-        return inBusyLock(busyLock, () -> mvPartitionStorages.finishRebalance(partitionId, mvPartitionStorage -> {
+    public CompletableFuture<Void> finishRebalancePartition(int partitionId, MvPartitionMeta partitionMeta) {
+        return busy(() -> mvPartitionStorages.finishRebalance(partitionId, mvPartitionStorage -> {
             try (WriteBatch writeBatch = new WriteBatch()) {
-                mvPartitionStorage.finishRebalance(writeBatch, lastAppliedIndex, lastAppliedTerm, groupConfig);
+                mvPartitionStorage.finishRebalance(writeBatch, partitionMeta);
 
                 indexes.finishRebalance(partitionId);
 
@@ -353,7 +367,7 @@ public class RocksDbTableStorage implements MvTableStorage {
 
     @Override
     public CompletableFuture<Void> clearPartition(int partitionId) {
-        return inBusyLock(busyLock, () -> mvPartitionStorages.clear(partitionId, mvPartitionStorage -> {
+        return busy(() -> mvPartitionStorages.clear(partitionId, mvPartitionStorage -> {
             List<AbstractRocksDbIndexStorage> indexStorages = indexes.getAllStorages(partitionId).collect(toList());
 
             try (WriteBatch writeBatch = new WriteBatch()) {
@@ -367,7 +381,9 @@ public class RocksDbTableStorage implements MvTableStorage {
 
                 return nullCompletedFuture();
             } catch (RocksDBException e) {
-                throw new StorageException("Error when trying to cleanup storage: [{}]", e, mvPartitionStorage.createStorageInfo());
+                throw new IgniteRocksDbException(
+                        String.format("Error when trying to cleanup storage: [%s]", mvPartitionStorage.createStorageInfo()), e
+                );
             } finally {
                 mvPartitionStorage.finishCleanup();
 
@@ -384,14 +400,18 @@ public class RocksDbTableStorage implements MvTableStorage {
     }
 
     private void checkPartitionExists(int partitionId) {
-        if (mvPartitionStorages.get(partitionId) == null) {
+        if (!partitionExists(partitionId)) {
             throw new StorageException(createMissingMvPartitionErrorMessage(partitionId));
         }
     }
 
+    private boolean partitionExists(int partitionId) {
+        return mvPartitionStorages.get(partitionId) != null;
+    }
+
     @Override
     public @Nullable IndexStorage getIndex(int partitionId, int indexId) {
-        return inBusyLock(busyLock, () -> {
+        return busy(() -> {
             checkPartitionExists(partitionId);
 
             return indexes.getIndex(partitionId, indexId);
@@ -401,5 +421,33 @@ public class RocksDbTableStorage implements MvTableStorage {
     @Override
     public StorageTableDescriptor getTableDescriptor() {
         return tableDescriptor;
+    }
+
+    private <V> V busy(Supplier<V> supplier) {
+        if (!busyLock.enterBusy()) {
+            throwExceptionDependingOnStorageState(state.get(), createStorageInfo());
+        }
+
+        try {
+            return supplier.get();
+        } finally {
+            busyLock.leaveBusy();
+        }
+    }
+
+    private void busy(Runnable action) {
+        if (!busyLock.enterBusy()) {
+            throwExceptionDependingOnStorageState(state.get(), createStorageInfo());
+        }
+
+        try {
+            action.run();
+        } finally {
+            busyLock.leaveBusy();
+        }
+    }
+
+    private String createStorageInfo() {
+        return IgniteStringFormatter.format("tableId={}", getTableId());
     }
 }
